[BUGFIX] Prevent CSP module showing link when EXT:lowlevel is missing
With #105692 a cross-reference link was added that guides people
to the "Configuration" backend module section for the CSP
configuration.

The used route "system_config" however is set via EXT:lowlevel,
which is not necessarily installed when the CSP backend module
is used.

Unfortunately, using a `<f:be.link>` with a missing route
throws an exception `RouteNotFoundException` that will
crash a whole backend module. This might be the proper
behavior for many cases, to prevent non-working links
in links used for JavaScript or form actions.

This patch checks for existence of EXT:lowlevel and
changes the intro text accordingly to hint at the
missing extension.

Resolves: #105988
Related: #105692
Releases: main, 13.4
Change-Id: I419940d023a1d4b817c36fda3a9952b463db7621
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87885
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>