[BUGFIX] Fix extbase persistence with wrong localized UID scheme
Extbase `<f:form>` editing with relations to localized records
utilizes a `[defaultLanguageRecordUid]_[localizedRecordUid]` identity
scheme.

This however cannot be resolved by the TypeConverter for persistence.

This patch spots identifiers created like that and only returns
their defaultLanguageRecordUid, which is the UID that should be
persisted so that overlays can work as currently implemented.

Resolves: #105319
Related: #59992
Releases: main, 13.4, 12.4
Change-Id: I42cb07e39a97f8fe668a910dd3f6315624342596
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88563
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>