[TASK] Remove workspaces BE module cache
The workspace BE module uses a cache that is more
harmful than good: The to be rendered data sets are
cached away, aiming to reduce query load when the
module is called over and over again for the same
page and "filter set". The cache solution has been
added more than ten years ago and is nowadays
highly questionable.

The cache collides with the general practice that
backend modules shouldn't cache such data since
good cache eviction is pretty hard to achieve in
all cases in backend modules.

Even though this cache *may* speed up repeated
clicks on the module, the cache is usually not
very effective and is also flushed often leading
to additional queries whenever DataHandler persists
something.

I've personally also seen stalled data being rendered
in the module and wondered about it, but I can not
reproduce an issue anymore, even though I'm pretty
sure it still existed. Still, the patch targets only
main/v14 since I can not pinpoint a bug directly.

The patch fully removes the workspace BE module cache
and relies on fresh data being fetched from database
each time. This avoids an entire layer of possible
bugs at the cost of a slower rendering on repeated
calls to the module for the same page. In case the
module really suffers from database load, we should
rather look on how to reduce this load. Editors will
most likely not feel a difference since the cache
often did not kick in anyway due to the repeated
flushes on data change.

Resolves: #106019
Releases: main
Change-Id: I9aac158f5f6bd47a11858e7049411c1bbf7394ca
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87932
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>