[TASK] Drop unneeded parameters from `Clipboard->confirmMsgText()`
The following parameters have been removed from
`Clipboard->confirmMsgText()`:

* `$selectedElements`
* `$columnLabel`

The parameter `$selectedElements` is whack as it contains the items in
the clipboard, passed from the outside. We can get the items from the
clipboard directly, as we have the necessary information available.

The parameter `$columnLabel` apparently was used in the page module
context but is unused. It is removed along with the according language
label.

Resolves: #106032
Releases: main
Change-Id: I410ddf666c69b2c503cde285400cab63e392f99f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87942
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>