[BUGFIX] Clarify searchable TCA fields
LiveSearch and List modules' SearchTermRestriction components
for each area use a different (static) list of TCA
types which are "searchable".

This change clarifies the list to be built in the Schema API
without keeping a static list everywhere.

This also resolves inconsitency and properly
validates the given "searchFields" list.

Resolves: #106040
Releases: main, 13.4
Change-Id: I2207ba802215eb019ce35ead900a2ed49fd2cd6a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87951
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>