[TASK] Add type hints in BackendUtility for some methods
Starting with `getQueryBuilderForTable()` and `getConnectionForTable()`,
type hints and return types are added to the functions and their
consumers. This is only done in cases the change is considered safe,
thus where a type error would be raised by subsequential method calls
that have the type already annotated.

As a drive by improvement, the mentioned helper methods are now used
consistently within the class and always called via `self::`, because
the class provides a static context and therefor does not require
late static binding.

Resolves: #106048
Releases: main, 13.4
Change-Id: I539d7b5102fc597d7bae149006c62e484aef0177
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87973
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>