[BUGFIX] Handle file upload validation for empty argument
If an extbase domain model only has one property of the type
`FileReference` or `ObjectStorage<FileReference>`, the extbase
`RequestBuilder` does not create an argument, if no file upload
is given in frontend context. The reason for this is, that an
empty input type=file field is not included in POST data.

Extbase file upload handling will validate properties with
a configured file upload by fetching the property value from
the argument. This validation however fails, when the argument
value is `null`, because `ObjectAccess::getPropertyPath` is
used to retrieve the property from a `null` value.

With this change, it is ensured, that the property value
is only resolved from the argument, if the argument is
not `null`. If the argument is `null`, the property value
is set to `null` too, ensuring that a file upload
property is correctly validated, if it is defined as
required.

Resolves: #105996
Releases: main, 13.4
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I7fce04954e820fc5dbce2b11b4c1f56029f163b1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87976
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>