[BUGFIX] Handle invalid site sets as a frontend rendering error
Sites with missing site sets are now detected during frontend
rendering and transformed into a page access failure in order
for sites to only be rendered if all sets are loaded.

This behaviour is more strict than before, but added to ensure
that sites are not silently rendered in degraded state without
ever noticing this issue (soft-500).

Resolves: #105516
Releases: main, 13.4
Change-Id: If3d2e43e41e70ddbff364d5aa8d28a6f73aed05e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88084
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>