[BUGFIX] Fix site set label loading on windows
The site set labels are loaded via an absolute
file path. That means a string like
LLL:C:/web/typo3/sysext/seo/Configuration/Sets/Sitemap/labels.xlf:label
is passed to the language API, where three colons are included because
of drives on windows (`C:/`).

Adapt the LLL string parser to be able to parse this kind of label
and pick the language from the part after the last colon, instead of
omitting the entire string after the second colon.

Also remove a previous workaround which was added to circumvent
the same parsing issue for the colon of the `EXT:`-prefix, which
is no longer needed with this fix.

Resolves: #106115
Related: #105713
Releases: main, 13.4
Change-Id: I78f9609eefb67ae0dab6931121a738cb3ba19346
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88077
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>