[!!!][TASK] Remove DataHandler->storeLogMessages
DataHandler logging can be fully disabled by setting
property enableLogging to false. This can be used during
bulk inserts and is used for instance by ext:impexp.

The second property storeLogMessages is of much less use.
It states: "Disabling will just save some database-space
as the default messages are not saved". This is deemed
rather useless in 2025 database engines. It is also
never used within TYPO3 core itself. It is also unused
when looking at TER extensions.

In general, we want to either have full log rows or none
at all, but not incomplete ones since this especially
collides with data tuples used in message placeholders.

The patch removes the property and its handling.

Resolves: #106118
Releases: main
Change-Id: I0af34b776a5e312a0c2e72db8a971bb318510b7b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88086
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>