[BUGFIX] Improve DH logging at row insert
DataHandler->insertDB() log handling is weird: It depends on
$newVersion that is partially derived from record state, and
partially derived from table tcaSchema->isWorkspaceAware().

The intention is to write log messages wether a new live
record has been added, or an new workspace overlay has been
added.

The patch simplifies the logic by looking at the incoming
record and removing the dependency to methods
getRecordPropertiesFromRow() and getRecordProperties().

There is one important side effect of this change: Current
log messages tried to state the 'record title'. This was done
by the internals of getRecordPropertiesFromRow() and
getRecordProperties() which trigger BU::getRecordTitle().
This is in general not a good idea. Log messages should
mainly refer to table:uid only since uids are the only thing
that don't change (in contrast to TCA derived titles).
Next, BackendUtility::getRecordTitle() can trigger a lot
of expensive sub calculations that should not be done within
DataHandler. Last, calculated record title is often empty,
leading to no useful information in logs.

This change reduces the DB load slightly since the record
in question is less often selected.

Further patches will continue improving DataHandler log
messages and there is potential to improve ext:belog with
good filter capabilities.

Note the v13 backport of this change will keep the existing
of the `@internal` instertDB() method, which is kept public
in this branch.

Resolves: #81727
Releases: main, 13.4
Change-Id: I2fc0ecd14ef86eebc04fd623890904bbb7222bd0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88091
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>