[BUGFIX] Show all file references even when one is missing
ElementInformationController shows links to resolved entries
of `sys_refindex`, which can be links/records but also files.

In case of a `sys_file_reference` entry, the code stopped
the whole execution workflow and did not process any
further elements.

This patch allows the foreach() loop to continue in case
of missing file references.

Resolves: #106139
Releases: main, 13.4, 12.4
Change-Id: I2b00e7624a4d961275f339fe0a222ed7325cabcc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88164
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Andreas Kienast <akienast@scripting-base.de>