[TASK] Use ForwardResponse instead of redirect in indexed search
When TypoScript config is missing for EXT:indexed_search,
the extbase action handling uses a redirect to
`noTypoScriptAction`, which changes the URI of a request
to point to that action.

When fixing the TypoScript configuration, a page reload
will thus still show the same error, resulting in a bad
developer experience.

By using a ForwardResponse() instead, the same output
can be displayed without a change of the URI.

Resolves: #106130
Releases: main, 13.4
Change-Id: Ib50792ca4f311c46859c55f2a629ceb2db218517
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88138
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>