[TASK] Allow reloading single record list tables via AJAX
This commit introduces a new endpoint that allows to reload single
tables in the record list view.

The first usage of this endpoint is when a record gets deleted. There
was always a chance that not everything was cleaned up properly,
therefore the tables are now fully reloaded.

This also lays ground for further UX improvements in the record list,
e.g. reloading only the affected table when the displayed columns are
changed via "Show columns".

A new CSS overlay is added to visually block the UI during the load
process, similar to the FormEngine overlay.

Resolves: #106241
Releases: main
Change-Id: Ic4124b02aa877722c422a83617ff0353fde40958
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88196
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>