[TASK] Remove 'center' and 'font' tags in RTE processing
The HTML tags 'center' and 'font' are deprecated. The CKEDitor
RTE by default utilizes '<span style="...">` tags to define
fonts and alignment.

With this change, the tags are added to the
`processing.HTLMparser_db.removeTags` configuration option.

This will no longer allow to store these two HTML tags in
database fields where the configuration is active, and
remove the tags from frontend output when the parseFunc
is active. A "important" ReST file describes this change
of behaviour, which is considered non-breaking in terms
of API, as it is a default configuration change only.

Tests are adapted for the new behavior.

Resolves: #106192
Related: #104839
Releases: main
Change-Id: I397261adb30d9630a19856f5f8c6f88bb264b72f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88256
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Wand <oliver.wand@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>