[TASK] Merge workspace hook moveRecord() into DataHandler
The core DataHandler handles lots of workspace related
details already, but still dispatches some handling over
to the workspaces DataHandlerHook. This makes refactoring
and bug fixing way more complicated than it should be since
the hook arguments are often rather unfortunate and can
not keep up with main DataHandler changes.

The patch is a straight inlining of the moveRecord()
workspaces hook handling directly into DataHandler: The main
method is merged into moveRecord(), three helper methods are
moved over as is, except mostly calls to $dataHandler, which
is now $this. The change refrains from adding excessive
refactoring even of obvious details to be easily reviewable.

This allows us to protect various methods that are now "free"
for signature changes, especially moveRecord_raw() is important
in this case. Note the v13 backport will not contain the method
visibility changes.

Resolves: #106224
Related: #106220
Releases: main, 13.4
Change-Id: I8bf56fef35e6699e1a837d21a4f76729babb8446
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88304
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>