[TASK] Simplify DataHandler->workspaceCannotEditRecord()
Method name and the "false" indicating "ok" are still not
cool, but the method internals can be simplified already.
The method is protected in v14 and now requires the full
record to be handed over instead of optionally the uid only.
All method consumers do that already.

Resolves: #106237
Related: #106236
Releases: main
Change-Id: Ifa497edeeac5160a11bf8de742b53be7e9595ce1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88327
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>