[BUGFIX] Provide browse menu key to the integrator
If using a menu of type "special = browse", it is really helpful
to know which of the given item names is retrieved. Those are derived
from the provided configuration "items" and needs to be added to the
retrieved database row as well.

The name is now added with the field name "ITEM_STATE" and can be used
when building a menu, e.g. with the menu dataprocessor.

Resolves: #105860
Releases: main, 13.4, 12.4
Change-Id: I99e91776d7cb66e3485d0c6aedd8ed4d9eed4e5e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88393
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>