[BUGFIX] Handle EXT:scheduler task deserialisation more gracefully
In case a scheduler task cannot be deserialized (for example,
tasks of extensions where an extension upgrade restructured central
object classes), also PHP TypeErrors are now caught and forwarded
with the specific `InvalidTaskException`, including the previous
exception/throwable.

A test is added for this scenario, with a manipulated serialized
string that would assign a mismatching type (string instead of array)
in the property `any`.

Resolves: #106243
Releases: main, 13.4, 12.4
Change-Id: I12039ba86f1e698285bc3ad9e90ed517b560889f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88361
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Lina Wolf <112@linawolf.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>