[BUGFIX] Handle empty, not-nullable passwords on record copy
By default, TCA `type=password` columns are set as `NOT NULL`
in the database (`config['nullable'] = false`).

When copying a record through FormEngine, an empty password
will be evaluated as `NULL` and the new record will inserted
as such, leading to an SQL failure "Column 'passwordfield' cannot
be null".

This patch checks on the nullable definition of a password field,
and in case of a NULL value converts it to an empty string, just
like when getting created through FormEngine (instead of copied).

This can be reproduced with EXT:styleguide when copying an
empty record of type "elements basic".

Resolves: #106283
Releases: main, 13.4
Change-Id: I6c1e37701efcfb188ddb6196eee7ef3f26285a7f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88564
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>