[TASK] Improve hints when no root page is found
When no root page is found, the Info box currently states
that no pages exist at all. However, this message also
appears when a root page exists but is invalid, which
may confuse integrators.

I updated the message to clarify this distinction and
added a link to the Getting Started tutorial on how to
create a root page.

Resolves: #106280
Releases: main, 13.4
Change-Id: I8a3b1bf4144a2f44a0ab0bc8cabb01a81e3428c8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88417
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Froemken <froemken@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Stefan Froemken <froemken@gmail.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>