Revert "[TASK] Allow reloading single record list tables via AJAX"
This reverts commit 4b08edee67c8d4bfae972f8a0587f8888ac54e94.

Reason for revert: This patch has some severe bugs. Also, it turned
out that reloading the record list tables via AJAX leads to loss of
user state which can only be solved in a meaningful way once the
record list itself is rendered as a web component.

Change-Id: Id9ecd1a4ebd76a2466e1eb96a8e249a8d63b4873
Resolves: #106286
Related: #106241
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88442
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>