[BUGFIX] Use the correct DI services scope in PackageActivationService
When activating an extension in TYPO3 classic mode,
a new DI container is built and made active for the PHP process
that runs the activation. Therefore it is required, that all services
that depend on the new container and TCA are fetched from the new container
and not already being injected into the service.

Also make sure, that when the container is reset, that
the TCA globals is also reset, so that the global array
matches the state in the initial service that is from then
on used by code executed after that point.

Resolves: #105342
Releases: main, 13.4
Change-Id: Iaa4f551bf5c3c3af081873ec6ae9979442cf8c4a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88512
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benni Mack <benni@typo3.org>